[chore] relocate context files
